corrected routes and controllers to create thought

+++ controllers/userController.js
@@ -1,12 +1,14 @@
 const { ObjectId } = require("mongoose").Types;
 // require the user model
-const { Username, User } = require("../models");
+const { User, Thought } = require("../models");
 
 module.exports = {
   // get all users
   async getUsers(req, res) {
     try {
-      const userData = await User.find();
+      const userData = await User.find()
+        .populate("thoughts")
+        .populate("friends");
 
       res.json(userData);
     } catch {
@@ -20,15 +22,15 @@ module.exports = {
     try {
       const user = await User.findOne({
         _id: req.params.userId,
-      }).select("-__v");
+      })
+        .populate("thoughts")
+        .populate("friends");
 
       if (!user) {
         return res.status(404).json({ message: "No user with that ID" });
       }
 
-      res.json({
-        user,
-      });
+      res.json(user);
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
@@ -64,19 +66,17 @@ module.exports = {
   },
 
   // delete a user
-  async deleteUser(req,res) {
+  async deleteUser(req, res) {
     try {
-        const user = await User.findOneAndDelete(
-            {_id: req.params.userId}
-        )
+      const user = await User.findOneAndDelete({ _id: req.params.userId });
 
-        if (!user) {
-            res.status(404).json({ message: "No user with that Id"})
-        }
+      if (!user) {
+        res.status(404).json({ message: "No user with that Id" });
+      }
 
-        res.json(user);
+      res.json(user);
     } catch (err) {
-        res.status(500).json(err);
+      res.status(500).json(err);
     }
-  }
+  },
 };
